create /create_application

+++ src/main/java/com/mr_syrus/credit/api/repository/ApplicationRepository.java
@@ -3,5 +3,5 @@ package com.mr_syrus.credit.api.repository;
 import com.mr_syrus.credit.api.entity.ApplicationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApplicationRepository extends JpaRepository<ApplicationEntity, Long> {
+public interface ApplicationRepository extends JpaRepository<ApplicationEntity, Integer> {
 }
